Flags superados: c, s, d, i.

+++ libft/libft.h
@@ -6,16 +6,14 @@
 /*   By: jgilaber <jgilaber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/01 21:40:06 by jgilaber          #+#    #+#             */
-/*   Updated: 2026/06/26 21:13:01 by jgilaber         ###   ########.fr       */
+/*   Updated: 2026/06/22 18:56:30 by jgilaber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
-# include <string.h>
 # include <unistd.h>
-# include <stdio.h>
 # include <stdlib.h>
 # include <stddef.h>
 
@@ -32,23 +30,23 @@ int		ft_isascii(int c);
 int		ft_isprint(int c);
 int		ft_toupper(int c);
 int		ft_tolower(int c);
-int	ft_strlen(const char *s);
-int	ft_strlcpy(char *dst, const char *src, int size);
+size_t	ft_strlen(const char *s);
+size_t	ft_strlcpy(char *dst, const char *src, size_t size);
 int		ft_strncmp(char *s1, char *s2, unsigned int n);
-int	ft_strlcat(char *dst, const char *src, int size);
-char	*ft_strnstr(const char *big, const char *little, int len);
+size_t	ft_strlcat(char *dst, const char *src, size_t size);
+char	*ft_strnstr(const char *big, const char *little, size_t len);
 int		ft_atoi(const char *nptr);
 char	*ft_strdup(const char *s);
 char	*ft_strchr(const char *str, int c);
 char	*ft_strrchr(const char *s, int c);
-void	ft_bzero(void *s, int n);
-void	*ft_memset(void *s, int c, int n);
-void	*ft_memcpy(void *dest, const void *src, int n);
-int		ft_memcmp(const void *s1, const void *s2, int n);
-void	*ft_memchr(const void *s, int c, int n);
-void	*ft_calloc(int nft_memb, int size);
-void	*ft_memmove(void *dest, const void *src, int n);
-char	*ft_substr(char const *s, unsigned int start, int len);
+void	ft_bzero(void *s, size_t n);
+void	*ft_memset(void *s, int c, size_t n);
+void	*ft_memcpy(void *dest, const void *src, size_t n);
+int		ft_memcmp(const void *s1, const void *s2, size_t n);
+void	*ft_memchr(const void *s, int c, size_t n);
+void	*ft_calloc(size_t nft_memb, size_t size);
+void	*ft_memmove(void *dest, const void *src, size_t n);
+char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_strtrim(char const *s1, char const *set);
 char	**ft_split(char const *s, char c);
@@ -68,6 +66,5 @@ void	ft_lstdelone(t_list *lst, void (*del)(void*));
 void	ft_lstclear(t_list **lst, void (*del)(void*));
 void	ft_lstiter(t_list *lst, void (*f)(void *));
 t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
-
 
 #endif
